Postgres: Remove old file unwrap logic. It's been slated for removal for a year & a half now
