+++ internal/databases/postgres/tar_interpreter.go
@@ -34,61 +34,6 @@ func NewFileTarInterpreter(
 		filesToUnwrap, newUnwrapResult(), createNewIncrementalFiles}
 }
 
-// TODO : unit tests
-func (tarInterpreter *FileTarInterpreter) unwrapRegularFileOld(fileReader io.Reader,
-	fileInfo *tar.Header,
-	targetPath string,
-	fsync bool) error {
-	if tarInterpreter.FilesToUnwrap != nil {
-		if _, ok := tarInterpreter.FilesToUnwrap[fileInfo.Name]; !ok {
-			// don't have to unwrap it this time
-			tracelog.DebugLogger.Printf("Don't have to unwrap '%s' this time\n", fileInfo.Name)
-			return nil
-		}
-	}
-	fileDescription, haveFileDescription := tarInterpreter.FilesMetadata.Files[fileInfo.Name]
-
-	// If this file is incremental we use it's base version from incremental path
-	if haveFileDescription && tarInterpreter.Sentinel.IsIncremental() && fileDescription.IsIncremented {
-		err := ApplyFileIncrement(targetPath, fileReader, tarInterpreter.createNewIncrementalFiles, fsync)
-		return errors.Wrapf(err, "Interpret: failed to apply increment for '%s'", targetPath)
-	}
-	err := PrepareDirs(fileInfo.Name, targetPath)
-	if err != nil {
-		return errors.Wrap(err, "Interpret: failed to create all directories")
-	}
-	file, err := os.OpenFile(targetPath, os.O_WRONLY|os.O_CREATE|os.O_TRUNC, 0666)
-	if err != nil {
-		return errors.Wrapf(err, "failed to create new file: '%s'", targetPath)
-	}
-
-	_, err = io.Copy(file, fileReader)
-	if err != nil {
-		err1 := file.Close()
-		if err1 != nil {
-			tracelog.ErrorLogger.Printf("Interpret: failed to close file '%s' because of error: %v", targetPath, err1)
-		}
-		err1 = os.Remove(targetPath)
-		if err1 != nil {
-			tracelog.ErrorLogger.Fatalf("Interpret: failed to remove file '%s' because of error: %v", targetPath, err1)
-		}
-		return errors.Wrap(err, "Interpret: copy failed")
-	}
-	defer utility.LoggedClose(file, "")
-
-	mode := os.FileMode(fileInfo.Mode)
-	if err = os.Chmod(file.Name(), mode); err != nil {
-		return errors.Wrap(err, "Interpret: chmod failed")
-	}
-
-	if fsync {
-		err = file.Sync()
-		return errors.Wrap(err, "Interpret: fsync failed")
-	}
-
-	return nil
-}
-
 // Interpret extracts a tar file to disk and creates needed directories.
 // Returns the first error encountered. Calls fsync after each file
 // is written successfully.
@@ -98,11 +43,7 @@ func (tarInterpreter *FileTarInterpreter) Interpret(fileReader io.Reader, fileIn
 	fsync := !viper.GetBool(internal.TarDisableFsyncSetting)
 	switch fileInfo.Typeflag {
 	case tar.TypeReg, tar.TypeRegA:
-		// temporary switch to determine if new unwrap logic should be used
-		if useNewUnwrapImplementation {
-			return tarInterpreter.unwrapRegularFileNew(fileReader, fileInfo, targetPath, fsync)
-		}
-		return tarInterpreter.unwrapRegularFileOld(fileReader, fileInfo, targetPath, fsync)
+		return tarInterpreter.unwrapRegularFile(fileReader, fileInfo, targetPath, fsync)
 	case tar.TypeDir:
 		err := os.MkdirAll(targetPath, 0755)
 		if err != nil {
@@ -132,4 +73,122 @@ func PrepareDirs(fileName string, targetPath string) error {
 	dir := strings.TrimSuffix(targetPath, base)
 	err := os.MkdirAll(dir, 0755)
 	return err
+}
+
+// TODO : unit tests
+func (tarInterpreter *FileTarInterpreter) unwrapRegularFile(fileReader io.Reader,
+	header *tar.Header,
+	targetPath string,
+	fsync bool) error {
+	if tarInterpreter.FilesToUnwrap != nil {
+		if _, ok := tarInterpreter.FilesToUnwrap[header.Name]; !ok {
+			// don't have to unwrap it this time
+			tracelog.DebugLogger.Printf("Don't have to unwrap '%s' this time\n", header.Name)
+			return nil
+		}
+	}
+	fileUnwrapper := getFileUnwrapper(tarInterpreter, header, targetPath)
+	localFile, isNewFile, err := getLocalFile(targetPath, header)
+	if err != nil {
+		return err
+	}
+	defer utility.LoggedClose(localFile, "")
+	defer utility.LoggedSync(localFile, "", fsync)
+	var unwrapResult *FileUnwrapResult
+	var unwrapError error
+	if isNewFile {
+		unwrapResult, unwrapError = fileUnwrapper.UnwrapNewFile(fileReader, header, localFile, fsync)
+	} else {
+		unwrapResult, unwrapError = fileUnwrapper.UnwrapExistingFile(fileReader, header, localFile, fsync)
+	}
+	if unwrapError != nil {
+		return unwrapError
+	}
+	tarInterpreter.AddFileUnwrapResult(unwrapResult, header.Name)
+	return nil
+}
+
+// get local file, create new if not existed
+func getLocalFile(targetPath string, header *tar.Header) (localFile *os.File, isNewFile bool, err error) {
+	if localFileInfo, _ := getLocalFileInfo(targetPath); localFileInfo != nil {
+		localFile, err = os.OpenFile(targetPath, os.O_RDWR, 0666)
+	} else {
+		localFile, err = createLocalFile(targetPath, header.Name)
+		isNewFile = true
+	}
+	return localFile, isNewFile, err
+}
+
+// get file unwrapper for file depending on backup type
+func getFileUnwrapper(tarInterpreter *FileTarInterpreter, header *tar.Header, targetPath string) IBackupFileUnwrapper {
+	fileDescription, haveFileDescription := tarInterpreter.FilesMetadata.Files[header.Name]
+	isIncremented := haveFileDescription && fileDescription.IsIncremented
+	var isPageFile bool
+	if localFileInfo, _ := getLocalFileInfo(targetPath); localFileInfo != nil {
+		isPageFile = isPagedFile(localFileInfo, targetPath)
+	}
+	options := &BackupFileOptions{isIncremented: isIncremented, isPageFile: isPageFile}
+
+	// todo: clearer catchup backup detection logic
+	isCatchup := tarInterpreter.createNewIncrementalFiles
+	if isCatchup {
+		return NewFileUnwrapper(CatchupBackupFileUnwrapper, options)
+	}
+	return NewFileUnwrapper(DefaultBackupFileUnwrapper, options)
+}
+
+// get file info by file path
+func getLocalFileInfo(targetPath string) (fileInfo os.FileInfo, err error) {
+	info, err := os.Stat(targetPath)
+	if os.IsNotExist(err) {
+		return nil, err
+	}
+	if info.IsDir() {
+		return nil, errors.New("requested file is directory. Aborting")
+	}
+	return info, nil
+}
+
+// create new local file on disk
+func createLocalFile(targetPath, name string) (*os.File, error) {
+	err := PrepareDirs(name, targetPath)
+	if err != nil {
+		return nil, errors.Wrap(err, "Interpret: failed to create all directories")
+	}
+	file, err := os.OpenFile(targetPath, os.O_WRONLY|os.O_CREATE|os.O_TRUNC, 0666)
+	if err != nil {
+		return nil, errors.Wrapf(err, "failed to create new file: '%s'", targetPath)
+	}
+	return file, nil
+}
+
+func (tarInterpreter *FileTarInterpreter) AddFileUnwrapResult(result *FileUnwrapResult, fileName string) {
+	switch result.FileUnwrapResultType {
+	case Skipped:
+		return
+	case Completed:
+		tarInterpreter.addToCompletedFiles(fileName)
+	case CreatedFromIncrement:
+		tarInterpreter.addToCreatedPageFiles(fileName, result.blockCount)
+	case WroteIncrementBlocks:
+		tarInterpreter.addToWrittenIncrementFiles(fileName, result.blockCount)
+	}
+}
+
+func (tarInterpreter *FileTarInterpreter) addToCompletedFiles(fileName string) {
+	tarInterpreter.UnwrapResult.completedFilesMutex.Lock()
+	tarInterpreter.UnwrapResult.completedFiles = append(tarInterpreter.UnwrapResult.completedFiles, fileName)
+	tarInterpreter.UnwrapResult.completedFilesMutex.Unlock()
+}
+
+func (tarInterpreter *FileTarInterpreter) addToCreatedPageFiles(fileName string, blocksToRestoreCount int64) {
+	tarInterpreter.UnwrapResult.createdPageFilesMutex.Lock()
+	tarInterpreter.UnwrapResult.createdPageFiles[fileName] = blocksToRestoreCount
+	tarInterpreter.UnwrapResult.createdPageFilesMutex.Unlock()
+}
+
+func (tarInterpreter *FileTarInterpreter) addToWrittenIncrementFiles(fileName string, writtenBlocksCount int64) {
+	tarInterpreter.UnwrapResult.writtenIncrementFilesMutex.Lock()
+	tarInterpreter.UnwrapResult.writtenIncrementFiles[fileName] = writtenBlocksCount
+	tarInterpreter.UnwrapResult.writtenIncrementFilesMutex.Unlock()
 }
